Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I don't understand the question? You can't have an *invalid prototype*, you can only have a prototype that's not in fast mode? If it's not in fast mode, the runtime should immediately make it fast after a miss because it was slow. That would avoid subsequent misses.

Just to have it recorded, the problem was because of a bug in the BranchIfPrototypeShouldbeFast and once the bug is fixed there are not many misses to runtime. Though not having this check didn't regress the performance when I tried speedometer locally. So for now, I will land this cl, since it simplifies the LoadIC_Uninitialized IC.

Patch-set: 3
Reviewer: Gerrit User 1145189 <1145189@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>