Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > > Patch Set 4:
> > > > 
> > > > Benedikt and I discussed this offline. We see two possible solutions:
> > > > 
> > > > 1) Enqueue the cleanup callbacks to the default MicrotaskQueue. This guarantees that the cleanup callbacks are run even after DetachGlobal is called. We need to check that the spec allows to run cleanup callbacks out-of-order w.r.t other microtasks (promise jobs).
> > > 
> > > Looks like the order doesn't matter per Sathya's comment, however, it's problematic to run user tasks on the default queue for several Blink features, such as Worklet, Cooperative Scheduling, and BFCache. They need to suppress user tasks from specific frame while they are working.
> > > Also, as we are deprecating to run JS functions after DetachGlobal [1], that is conflicting to the FinalizationGroup clean up.
> > > [1]: https://bugs.chromium.org/p/chromium/issues/detail?id=897927
> > > 
> > > > 
> > > > 2) Change the lifetime of MicrotaskQueue, so that it is alive and is processed as long as the Context is alive. I guess this is difficult to implement in Blink and is problematic for the HTML spec?
> > > 
> > > There's a similar on-going discussion for the promise cancellation on detached contexts [2].
> > > I don't have enough time to do it right now, but it's possible by moving everything to GCed area.
> > > [2]: https://github.com/whatwg/html/issues/4443
> > > 
> > > > 
> > > > Note that invoking the cleanup callbacks on DetachGlobal is not sufficient because GC can discover new cleared JSWeakRefs after DetachGlobal. To avoid memory leaks, we need to run the cleanup callback as long as the Context is alive. Once the Context dies, there will be no cleanup callbacks (because cleanup callbacks are enqueued for live FinalizationGroups and FinalizationGroups keep the Context alive).
> > > 
> > > OK, that makes sense.
> > > 
> > > Is it possible to distinguish the user-defined function to system ones? Does the clean up callback used by WebAssembly runs any user code? If yes and no, we may be able to drop all user task while keep running system tasks such as promise resolution and wasm clean up.
> > 
> > I don't see a good solution here. I'd prefer to avoid special-casing wasm callbacks.
> > Given that we want to deprecate JS functions after DetachGlobal, canceling cleanup callbacks seems like the most reasonable option here.
> > 
> > Is is possible for users to subscribe to an event related to DetachGlobal? That would be the place to eagerly cleanup resources associated with the current context.
> > 
> > Sathya, should we accept that cleanup callbacks are not invoked after DetachGlobal?
> 
> It seems to me that the answer to the first of these two questions would help answer the second. That is, if there's a signal authors can use to do their cleanup on detachment, it could be used to handle cases where WeakRef cleanup is going to be canceled.
> 
> But I'm still fuzzy on how long these detached contexts exist for; are we worried about leaks within those contexts, or not?

Upon detach we'll actually remove the global object from it's native context, and we'll guarantee that no code can run in there anymore. We want to clear out that context and anything it holds onto as eagerly as possible without interaction from the site. Basically a "force quit" on the native context.

The thing that Benedikt mentions is the part I'm providing under a flag: Disallow calling functions from a detached context entirely. This came from my comments that just clearing microtasks is weird, as we'd still be able to call async functions, but they would never proceed beyond await...

I'm not as clear on the details as Sathya and Adam are probably as to what is possible and what could possibly still leak from e.g., wasm. Since the iframe could have been same-origin as a native context that survives, we can't just randomly clear out objects. Only the global object can be cleared out since it's inaccessible (code can't run, so no direct access; and the global proxy is already attached to a different global object). That makes it seem to me that indeed, if the page isn't allowed to clean up before we navigate away, it will possibly leak memory indefinitely in objects that were leaked to frames that still survive.

While that does sound like a problem, the problem this is trying to solve is likely much worse in real-world current scenarios. As an interim solution it sounds ok to keep contexts alive that have FinalizationGroups; but likely this isn't a good long-term solution since pages that care about being a good citizen will become a bad citizen... 😕

Patch-set: 4
Reviewer: Gerrit User 1184114 <1184114@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>