[promise] Lookup the resolve property only once
In the PerformPromise{All, Race, AllSettled} operations, the resolve
property of the constructor is looked up only once. For the fast path,
where the constructor's resolve property is untainted, the resolve
function is set to undefined. Since undefined can't be a valid value
for the resolve function, we can switch on it (in CallResolve) to
directly call the PromiseResolve builtin. If the resolve property is
tainted, we do an observable property lookup, save this value, and
call this property later (in CallResolve).

Spec:
- https://github.com/tc39/ecma262/pull/1506
- https://github.com/tc39/proposal-promise-allSettled/pull/40

Bug: v8:9152
Change-Id: Icb36a90b5a244a67a729611c7b3315d2c29de6e9