Update patch set 1
Patch Set 1:

PTAL. Dumping it on this bot to see if it works. We might want a different bot though since the trybot uses dchecks. That might be to slow for a benchmark. We'd want a pure release builder.

So maybe we need a separate set of test on CI and builder on tryserver for this. That builder could then be added with a path matching on changes to run_perf.

Patch-set: 1
Reviewer: Gerrit User 1153089 <1153089@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1115956 <1115956@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false