[snapshot] Align internal snapshot data
When the snapshot blob is not aligned properly, loading it can cause a
crash on platforms such as arm.

This was exposed by a SIGBUS/BUS_ADRALN crash on arm when accessing
the blob_data symbol (declared as a byte array) through a reinterpret
cast to uintptr_t in an internal snapshot build.

Bug: v8:9171
Change-Id: I99b071dec3733416f2f01b58a770e30d8f2dcdf2