Update patch set 18
Patch Set 18:

I ran into an issue with this example:


class D extends class E {} {
  #d() {}
}

When calling `new D`, it throws an error.


1. Calling D, AddPrivateBrand D
2. Calling E as super, AddPrivateBrand E, and then AddPrivateBrand D because D is in
  outer scope of E - then throws an error because brand D is already added.

I am not entirely how to eliminate the second AddPrivateBrand for D. Maybe we could special case so that when building a super call, skip classes in the outer scope that are subclasses
of the current class?

It is not in the current patch set, because AddPrivateBrand is only called in
2 when the base class has something private and requires a member initializer.
However the current patch throws for this

class D extends class E {
  #e() {}
} {
  #d() {}
}
new D();

This is not an issue for the fields because we don't error for them when the fields are previously added.

Patch-set: 18
Reviewer: Gerrit User 1288325 <1288325@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0