Update patch set 2
Patch Set 2:

Thanks for the review, Jakob. Please see my replies inline.
> Patch Set 2:
> 
> Two identical strings by definition also have the same length, no? Why is the length comparison necessary? 
I was under the impression that since javascript strings are immutable and come with length as extra information instead of null-terminated it might be possible to have different strings of different lengths but pointing to the same memory. So when we compare such strings we could just look at their lengths, hence the comparison. I'm still learning about javascript strings, so please correct me if this assumption was wrong.
 
> Are comparisons against empty strings relevant in practice?
It triggered in two of the js unittests and a CDJS subtest in JetStream2.0 benchmark during key/value insertions into RB-Tree so I thought it might be relevant.
> 
> In general, do you have data on the performance impact of these changes? I would think that String::Flatten is very quick when the strings are flat already, so I'd be curious to learn how much this patch helps.
I am collecting the data on our devices and will update this thread when I get it.

Patch-set: 2
Reviewer: Gerrit User 1325086 <1325086@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>