Update patch set 2
Patch Set 2:

It's true that Strings in V8 are not null-terminated, but that's unrelated to what you're doing here. Using the * operator on a Handle<String> gives you the plain String object, and if two of those are equal, that means they are the same string (same length, same contents). This is true across all String types. (Note that the reverse implication does not hold: in general, Strings are not canonicalized, so it is well possible to have different String objects with the same length/contents. Only when comparing two "internalized" Strings it is guaranteed that they have equal contents if and only if they are the same object.)

Unit tests do not generally constitute evidence of practical relevance. We like to have tests even for slow code paths that are rarely used. The gold standard for practical relevance is when a well-known website or framework uses the pattern in question.

I'm looking forward to performance data on this patch!

Patch-set: 2
Reviewer: Gerrit User 1115935 <1115935@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>