Update patch set 4
Patch Set 4:

Sorry for the delay.

I'm not sure I agree with the motivation here. In my mind, this is only a relation between modules and memories and I don't see a need for a global engine mode. There is actually less coupling if we don't have a global mode. I.e. if a module and its instance's memory agree on whether bounds checks or guard regions are used, then this is good enough.

Patch-set: 4
Reviewer: Gerrit User 1119499 <1119499@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>