Update patch set 7
Patch Set 7:

I used to run d8 like this:

    d8 --wasm-tier-up --liftoff x.js --wasm-num-compilation-tasks=0 --wasm-lazy-compilation

Where x.js basically calls

    WebAssembly.compile(wasm_bytes)

This CL now causes it to crash with:

    # Fatal error in ../../src/wasm/module-compiler.cc, line 1955
    # Debug check failed: FLAG_wasm_lazy_compilation || (FLAG_asm_wasm_lazy_compilation && module->origin == kAsmJsOrigin) implies outstanding_baseline_functions_ == 0.

This seems to happen becuase `InitializeCompilationProgress` is called from `PrepareAndStartCompile`, but the `NativeModule` has not had `set_lazy_compilation` called on it, like is done in `CompileNativeModule`.

Is that use of d8 supposed to be supported?

Patch-set: 7
CC: Gerrit User 1334495 <1334495@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>