Update patch set 1
Patch Set 1:

I don't believe it's possible for the Module to be collected in that case. The Module's Context (effectively, the object representing the module's scope) holds a reference back to the Module, so it can't be collected as long as some closure in that module is alive.

Patch-set: 1
Reviewer: Gerrit User 1115851 <1115851@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>