Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I agree with Georgia's assessment that in this version the size of the deoptimization bailout would increase drastically (especially given the required push/pop sequence as well). The execution speed of the deoptimization bailouts is not so much a problem, but I think their code size does matter.
> 
> As for the suggestions that Jakob mentioned, my intuition is leaning towards passing the deopt id via the stack. Using an Isolate-wide slot (i.e. smi-root) sounds like it would work too given JavaScript's inherent single-threaded nature, but somehow feels more restrictive and less local.
> 
> I am also adding Jaro, because this will definitely also need sign-off from his side.


Thank you all for the insightful comments!!! I apologize for this very late reply.

I've been trying to implement your suggestion and pass the deopt_id via the stack. This is actually very simple to do. But I am still concerned with the problem of the size of the deoptimization bailout in some platforms.
For example, if we want to push the deopt_id on ARM64, we would have something like:

void TurboAssembler::CallForDeoptimization(Address target, int deopt_id) {
  UseScratchRegisterScope temps(this);
  Register temp = temps.AcquireX();

  movz(temp, deopt_id);
  Push(temp, padreg);

  IndirectCall(target, RelocInfo::OFF_HEAP_TARGET);
}

Here 'target' is the address of the GenerateDeoptimizationEntries() builtin, possibly embedded in the V8 executable. I don't know if it is possible to generate more efficient code, but this code would still be compiled into 6 instructions:

movz(temp, deopt_id) => movz
Push(temp, padreg)   => stp
IndirectCall(...)    => movz, movk, movk, blr

The problem is that the call to a far-away address may require 4 instructions. Instead, the current code relies on the fact that it can do a near-call with just one 'bl' instruction:

void TurboAssembler::CallForDeoptimization(Address target, int deopt_id) {
  BlockPoolsScope scope(this);
  NoRootArrayScope no_root_array(this);

  movz(x26, deopt_id);
  int64_t offset = static_cast<int64_t>(target) -
                   static_cast<int64_t>(options().code_range_start);
  DCHECK_EQ(offset % kInstrSize, 0);
  offset = offset / static_cast<int>(kInstrSize);
  DCHECK(IsNearCallOffset(offset));
  near_call(static_cast<int>(offset), RelocInfo::RUNTIME_ENTRY);
}

So even pushing the deopt_id onto the stack we would still increase the size of this code from 2 to 6 instructions, wasting 16 bytes for each thunk, which is probably not acceptable. 

I am really not an expert of ARM assembly, but it looks like the real problem here is that if we transform the stubs into builtins, the Call instructions targeting addresses in the embedded code necessarily take more space. 
Do you think there could a way to minimize this problem, or should we just abandon the idea of transforming DeoptimizationEntries into embedded builtins?

Patch-set: 3
Reviewer: Gerrit User 1318771 <1318771@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0