Update patch set 1
Patch Set 1:

This is my first Chromium patch, so I apologise if I've missed anything. I wasn't sure who to add as a reviewer, but I'm adding Michael, as I saw he has interacted with the bug.

I noticed that some projects under Chromium's third_party (like Polymer and Speedometer) include the same string, and in a few cases, that is as part of a test. Changing those projects felt wrong, so I haven't done so. I think in most cases it was just because they try to copy V8's errors, but it's possible in one or two cases their tests will break after this change. Mentioning just in case the side effects of this change need to be considered.

Patch-set: 1
Reviewer: Gerrit User 1344892 <1344892@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false