Update patch set 9
Patch Set 9:

> Patch Set 2: Commit-Queue+1
> 
> Georg, PTAL at the first CL to enable brokerization of the inlining.
> 
> Dan, Ross, that's FYI. Please note that running cctest with and without these change seems to produce the same results, when --trace-turbo/--trace-deopt and/or --enable-lazy-source-positions are enabled. For instance the following two runs result in failures both on ToT and with this change:
> 
> out.gn/.../cctest test-cpu-profiler/Inlining2 --enable-lazy-source-positions
> out.gn/.../cctest test-interpreter/InterpreterCollectSourcePositions_GenerateStackTrace --trace-deopt
> 
> Are those failures expected? Any comments are welcome!

The Inling2 test is still a WIP for lazy source positions.

The second test is testing what happens if you have lazy source positions and then generate a stack trace. If you enable trace-deopt then I think it will cause it to eagerly generate source positions and so the test can't function properly. I don't know that it makes sense to rework the test to detect this scenario and pass anyway as then it's possible it could be forced into such a state permanently making the test always pass regardless of whether lazy source positions actually work.

Patch-set: 9