Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> It seems a bit surprising to run tests for a third_party library as part of v8's unittests. Generally I think of third_party stuff as being tested upstream (we don't test, e.g., perfetto as part of our build).
> 
> Can you give some more background on it's important that this test run here, rather than in some upstream CI?

Yeah, we don't have an upstream CI, it's a small project. I try my best to run the test upstream, however, v8 has a fair number of different configurations that would be difficult to reproduce there (gcc vs. clang, c++11 vs 14, etc.), so I like it to make running the test part of the roll so we don't introduce broken stuff.

I've done a similar thing for Chromium, we also run the test there as part of the content_unittests. Running it both for v8 and for Chromium makes the rolls independent, which is good.

On a positive note, the test is fast, about 50ms end to end on my box.

I think the precedent for this scheme is gunit, for which we run tests in chromium and v8 and mini_chromium also, even though it probably does run upstream tests as well.

Thanks much for considering.

Patch-set: 11
Reviewer: Gerrit User 1280850 <1280850@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0