[error] add flag to retain the array stack trace
Users of postmortem debugging tools can leverage the stack trace array
in a Error object to inspect the stack and state of an application which
crashed asyncronously, either from within a Promise, an async function
or from a rethrown error. Retaining the array can introduce memory
overhead, so we kept the default as it was (clean the stack after
.stack is accessed from JavaScript), and we added a flag for users who
want to pay the memory overhead in exchange for a better postmortem
experience.

R=yangguo@google.com

Change-Id: Iae45f87531e88cb430ed75d5e91c3d0635a73c71