Update patch set 3
Patch Set 3:

When relanding with a fix, upload the original patch first and then the fix as the next patchset so it's easy to see the diff.

It's hard to know what the already reviewed code is and what the new fix is with a big CL like this.

Patch-set: 3
Reviewer: Gerrit User 1158954 <1158954@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>