[tracing] Add support for pending trace events to Perfetto
"Complete" events (where phase='X') have a begin/end part managed by a
scoped object on the stack. When the object goes out of scope, the
trace event is updated with the duration of the entire event.

Perfetto does not support this directly as it requires first logging
the event in the buffer, and then later updating it, which is not
compatible with the append-only protobuf library used to write into
perfetto buffers.

This CL adds a thread-local stack of pending trace events. When we get
the start of a two-part trace event we don't immediately write it into
the buffer, we instead add it to the pending stack. When we get the end
of the trace event and need to update the duration, we remove it from
the stack and write the entire event to the trace buffer. This is the
same approach that Chrome uses to solve the same problem.

Introduces a TempTraceRecord struct which holds the trace event details
in the thread-local stack.

Cq-Include-Trybots: luci.v8.try:v8_linux64_perfetto_dbg_ng
Bug: v8:8339
Change-Id: I1a581672b62dd206fd84bc60c19ad97e4a0bd3b2