[map] In FindElementsKindTransitionedMap only check for required element kinds
FindElementsKindTransitionedMap checks if we could transition from
the source map to one of the target maps without requiring any instance
rewriting. It does this by replaying all the property transitions of the
source map on each elements kind map and seeing if it needs a instance
rewrite. Since we already know the elements kind of the target maps, we can
avoid doing this for element kinds that are not in target map.

Change-Id: Ief9ba89992a411535a0335c3b67221666647f55e