[Android] Don't apply offset to library start address for libs mapped from APK.
On Android, we mmap the .so directly from the APK, therefore the file offset
in this case refers to the offset of the .so in the APK, not the offset in
the .so file itself. As such, moving the start position of the library based
on this gives incorrect results when later symbolizing the addresses based on
the .so.

Change-Id: I16f78c91106dc85334a8b0034c96b19442af7b05