Update patch set 1
Patch Set 1:

I was trying to understand map updates a bit better and trying to trace the origin of the strange boilerplate values we discussed last week. In the process I adapted the debug printing of generalizations to also show constness, as we discussed a while back.

PTAL, for a couple of the call sites I'm not sure which is supposed to be the old/new constness.

Patch-set: 1
Reviewer: Gerrit User 1345174 <1345174@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false