[torque-ls] Add a simple version checking mechanism
This CL adds a version header with a constexpr constant that signifies
a torque compiler/language server version. On startup, the language
server reads in the checked out header, parses the version number
and notifies the LSP client if the current binary version does not
match the version found in the header.

Bug: v8:8880
Change-Id: I45b37221be04c5385b42196bccd54205f5430228